[nobank] Add return flag in fetchXXX method(), so we can know whether app call network request. 1. Update unit test, use ShadowLooper.runUiThreadTasksIncludingDelayedTasks(); to wait handler work done. 2. move some interface classes to manager.xxxBasic package.
