+++ NobankCore/noBank/src/main/java/com/hunk/nobank/extension/network/MyNetworkHandler.java
@@ -5,7 +5,7 @@ import android.os.Handler;
 import android.os.Message;
 
 import com.hunk.nobank.contract.RealResp;
-import com.hunk.nobank.manager.ManagerListener;
+import com.hunk.nobank.manager.dataBasic.ManagerListener;
 import com.hunk.nobank.model.Cacheable;
 import com.hunk.nobank.util.Logging;
 import com.squareup.okhttp.Callback;
@@ -103,7 +103,7 @@ public class MyNetworkHandler extends NetworkHandler {
         return realResp;
     }
 
-    static class Wrapper {
+    public static class Wrapper {
         public final String managerId;
         public final ManagerListener listener;
         public final Object realResp;
