+++ NobankCore/noBank/src/main/java/com/hunk/nobank/extension/network/NetworkHandler.java
@@ -14,7 +14,7 @@ import com.google.gson.Gson;
 import com.hunk.nobank.contract.ContractGson;
 import com.hunk.nobank.contract.RealReq;
 import com.hunk.nobank.contract.RealResp;
-import com.hunk.nobank.manager.ManagerListener;
+import com.hunk.nobank.manager.dataBasic.ManagerListener;
 import com.hunk.nobank.util.Logging;
 
 import java.io.UnsupportedEncodingException;
