+++ NobankCore/noBank/src/main/java/com/hunk/nobank/manager/ScreenFlowManager.java
@@ -1,4 +1,6 @@
-package com.hunk.nobank.manager.flow;
+package com.hunk.nobank.manager;
+
+import com.hunk.nobank.manager.flowBasic.ScreenFlow;
 
 /**
  * @author HunkDeng
