+++ NobankCore/noBank/src/main/java/com/hunk/nobank/manager/dataBasic/DataManager.java
@@ -1,4 +1,4 @@
-package com.hunk.nobank.manager;
+package com.hunk.nobank.manager.dataBasic;
 
 public abstract class DataManager {
     public abstract String getManagerId();
