+++ NobankCore/noBank/src/main/java/com/hunk/nobank/manager/dataBasic/ManagerListener.java
@@ -1,4 +1,4 @@
-package com.hunk.nobank.manager;
+package com.hunk.nobank.manager.dataBasic;
 
 public interface ManagerListener {
     void success(String managerId, String messageId, Object data);
