+++ NobankCore/noBank/src/main/java/com/hunk/nobank/manager/dataBasic/ViewManagerListener.java
@@ -1,4 +1,4 @@
-package com.hunk.nobank.manager;
+package com.hunk.nobank.manager.dataBasic;
 
 import java.lang.ref.WeakReference;
